Added template based styles and hooked _Person.vue to contentLoader
I added default color styles to the _Person.vue page and more return categories to the PersonCard to match the design but the data is not present in the API

Created a list of data we will need from the API to correctly display the PersonCards the PersonInfo and the map.

Got the PersonInfo page loading with the contentLoder and started to look into the data that will be needed by the ModMap component. Had to comment out existing link to mod-map