Resolved all content spacing issues with or without all the data coming in.
All spacing will react appropriately even if data is missing. This includes images. Didn't address any issues from 9/30 client review meeting yet. This work was to have a good base for that. Still need to resolve issues for the Room page. Might be refactoring that design to match Campus and Building